+++ deltafi-api/lib/deltafi/metrics.rb
@@ -27,7 +27,7 @@ module Deltafi
     extend self
 
     def self.graphite(query, raw: false)
-      graphite_url = "#{ENV['DELTAFI_GRAPHITE_URL'] || DF.cached_system_properties['graphite.url'] || 'http://deltafi-graphite:8080'}/render"
+      graphite_url = "#{ENV['DELTAFI_GRAPHITE_URL'] || 'http://deltafi-graphite:8080'}/render"
 
       if raw
         HTTParty.post(graphite_url, query: query).to_s
@@ -42,7 +42,7 @@ module Deltafi
       end
     end
 
-    GRAPHITE_METRIC_URL = ENV['DELTAFI_GRAPHITE_METRIC_URL'] || DF.cached_system_properties['graphite.metric-url'] || 'tcp://deltafi-graphite:2003'
+    GRAPHITE_METRIC_URL = ENV['DELTAFI_GRAPHITE_METRIC_URL'] || 'tcp://deltafi-graphite:2003'
     @@graphite_client = GraphiteAPI.new graphite: GRAPHITE_METRIC_URL, slice: 10
 
     def self.record_metric(name:, value:, tags: {}, prefix: nil, gauge: false)